+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ### Added
 
 - `scoped` property creator for scoped interactors
+- `only` method so nested interactors can break out of parent chains
 
 ### Changed